requirements.txt
discord.py
easy_pil
requests